# PERF: construct DataFrame with string array and dtype=str