@jorisvandenbossche 2020-09-17
Nice! Add a benchmark for it? 
And is this speed-up only for constructing from an array, or also eg from dict `{'col': x}` ?