@topper-123 2020-09-17
> And is this speed-up only for constructing from an array, or also eg from dict {'col': x} ?

This PR just affects the path for multi-dimensional arrays, while `{'col': x}` takes the 1-dim path, so not affected by this. The one dim-path was however helped by #36317.