@topper-123 2020-09-17
> Nice! Add a benchmark for it?

Yeah, I'll add it after the travis checks work again.